refactor: move web layer from adapter/controller into infra/web/gin
Collapse the adapter/controller and adapter/gateway trees into
infra/web/gin (handlers, middleware, presenter, router, server).
